Preparation of Reference Files
In order to run the STICR analysis pipelines for both plasmid and experimental (i.e. scRNA-seq) analysis, you will need to first create bowtie references that the pipeline uses to map the STICR fragments. Do this using standard bowtie reference generation as described in their documentation and place the references in the path that you will reference as INDEX_PATH in the SCTIR scripts. You also need to download and place the file “E_Selections.fa” in the same path.